Remove deprecated components: CourseSelector, Navigator, ScheduleTable, and related files to streamline the project structure.
